Add checkout
Background
==========
Implement MVP for a checkout feature. The Store loads products from a
json file. In the checkout a user can add products to the basket by
using the `Checkout.scan/2` function and calculate the total for the
products in the basket. The `Checkout.total/1` function applies
discounts if the basket qualifies for any.

Additional Notes
================
At the moment every time a product needs to be found by code the whole
json file needs to be parsed, an array of products needs to be build and
the array of product needs to be traversed every time.

This approach is far from optimal. We should parse the json config only
once and cache the products. I'm leaving this for a future iteration.